+++ templates/forgot_password.html
@@ -51,18 +51,27 @@
                     
                     const data = await response.json();
                     
-                    // Success branch, shows PIN and gives evaluators time to copy it for the next screen
+                    // Success branch
                     if (data.status === 'success') {
                         const pin = data?.debug_info?.pin
                             ? String(data.debug_info.pin)
-                            : (data.message || '').replace('Reset PIN:', '').trim();
+                            : null;
+                        if (pin) {
+                            // Show PIN and link
+                            document.getElementById('message').innerHTML = 
+                                `<strong>Reset PIN:</strong> ${pin}<br><a href="/reset-password">Continue to Reset Password</a>`;
+                        } else {
+                            // Hardened, no PIN shown
+                            document.getElementById('message').textContent = data.message;
+                            setTimeout (() => {
+                                window.location.href = '/reset-password';
+                            }, 2000);
+                        }
 
-                        document.getElementById('message').innerHTML =
-                            `<strong>Reset PIN:</strong> ${pin}<br><a href="/reset-password">Continue to Reset Password</a>`;
                         document.getElementById('message').style.color = 'green';
                         document.getElementById('message').className = 'success';
                     } else {
-                        document.getElementById('message').innerHTML = data.message;
+                        document.getElementById('message').textContent = data.message;
                         document.getElementById('message').style.color = 'red';
                         document.getElementById('message').className = 'error';
                     }
